no need for GetNoteSkinActor() to be global
Similar to 081c350293c904ff6f1ae6292357fd8dfec93650, this takes existing functionality that was wrapped up in a function with global scope (defined in the ./Scripts directory) and moves it elsewhere so that it is no longer global.

In this case, I've moved the GetNoteSkinActor helper function out of SL-Helpers.lua and put it into BGAnimations/_modules/NoteSkinPreview.lua

Similar to 92bcf177482c760536b2fafdc9b57149e09ad54a, I have no practical means of assessing whether this is better or worse for SL's memory footprint, just hope.