start to support split BPMs in ScreenPlayerOptions
.ssc charts support individual stepcharts having their own BPM values.  This is in contrast to .sm files where a single set of BPMs applies to every stepchart in the file.

SM5 has supported "slit timing" like this for quite a while, and Simply Love has pretty much just... not.  Simply Love's layout was designed for an engine that didn't have to worry about this.

Until now, in cases where I needed to display the current BPM value I would use the current player's BPM if one player was joined or the MasterPlayerNumber's BPM is both were joined.  Most of the time this was fine because, to be honest, very few songs use split timing.  It quietly nagged me, though.

So, in my recent commits 5c20a596cd41b46459d4cbde5e4ff414876180a3 and 0e032e099b75bdd88ebf5c37b5c851f2eb12e937 I added support for split BPMs to SelectMusic.

This commit adds mostly-working support to ScreenPlayerOptions.

Part of making this work was creating a custom OptionRow for choosing a difficulty (beginner, easy, medium, hard, expert, edit).  It was previously handled by the engine.  I needed to be able to broadcast when the selected difficulty changed and update the necessary scroll-speed displays on ScreenPlayerOptions in realtime if the entire BPM range for that player had suddenly changed.

In using Ace For Aces and one custom chart of my own as test cases, I've played around with having both players joined, switching between [X, C, M] mods, changing music rate, and changing difficulties.  This part of it seems to work well.

There are currently two bugs:

1. The help text in the gutter for the MusicRate OptionRow doesn't yet support split BPMs.  There's not enough space there to display 2 completely different BPM ranges simultaneously.  I have a plan to move this information elsewhere in the screen but haven't gotten to it yet.  It shouldn't be too hard.

2. If you change your difficulty on ScreenPlayerOptions using the new custom OptionRow, your difficulty will apply correctly and you'll play the stepchart you wanted in Gameplay, but your difficulty will then reset when you return to SelectMusic.

My testing so far suggests that when you get back to SelectMusic, the difficulties are set correctly at the time the InitCommand for the main ActorFrame of SelectMusic overlay is called, but they have been reset by the time of OnCommand.

I'll need to keep digging into this, but wanted to commit what I have for now.