make a GetPlayableTrails function
Since SL is now handling the OptionRow for Difficulty, it's up to me to determine which stepcharts in an sm or ssc file are playable given the current state of things.  If we're in dance versus, we shouldn't see dance double charts appear in the list, or pump single charts, or etc.

The SongUtil table exposed to Lua by the engine has a GetPlayableSteps() function we can use for normal gameplay, but there is no analogous CourseUtil.GetPlayableTrails() or anything similar that I can see for CourseMode.

This commit adds such a function to the top of SL-PlayerOptions.lua.  It's local to the file and now in use.  It seems to work okay.

The difficulty reset bug described in fc14746bb56095c0118ab8e81d73ac9681a013b4 is still an open issue.