fix EvalSummary difficulty and meter in CourseMode
While testing CourseMode, I noticed that EvaluationSummary wasn't displaying the appropriate difficulty meter for each course.

I'd been using the "Stamina Safari - Expert 15s" course from ITGAlex's Stamina Safari as my test case.  It's a course full of 15s that was given a difficulty meter of 17, overall.

#COURSE:Stamina Safari - Expert 15s;
#METER:Regular:17;

This was displaying on ScreenEvaluationSummary as 15.  It seemed my previous code was using the difficulty and difficulty meter from the last song in the course.

It should be fixed now.