"Difficulty" → "Stepchart"
fc14746bb56095c0118ab8e81d73ac9681a013b4 introduced a themeside OptionRow for
handling stepchart selection on ScreenPlayerOptions to facilitate support for
split bpms.

At that time, I failed to notice that the help text at the bottom of the screen
was no longer appearing for that OptionRow.  I'd named that OptionRow
"Difficulty" when the engine had been referring to it as "Steps" and forgotten
to update the language files accordingly.

I'm settling on "Stepchart" for now.