remove unnecessary MusicRate compensation
1da0e92c12373c35cb7928728d7868e009f56c6b fixed (and cleaned up) the code for
switching SpeedModType (for example, from X to C) when MusicRate modifiers were
in effect.

It seems like the code added then is now unnecessary as MusicRate is handled
("compensated for") by the global GetDisplayBPMs() helper function.

To be honest, I never play with MusicRate and I'm not 100% sure what the
desired behavior is supposed to be, but this feels like it makes more sense.  I
hope.  I can revert this if not.