more consistent string formatting in SelectProfile
If an old version of Simply Love had saved SpeedMod=1.50000000001234 to a
player's Simply Love UserPrefs.ini, that would appear as such on
ScreenSelectProfile.

I can't control what is out in the wild in player profiles, but I can try to
format strings in the theme more consistently going forward.