show profile HighScores in Eval's Pane3
ScreenEvaluation's Pane3 has traditionally shown the top 10 machine HighScores for the stepchart that was just played.

This commit changes Pane3 to check if a profile (local or USB) is in use, and show

• top 7 machine HighScores, and then
• top 3 personal HighScores

if so.  If a profile isn't in use, the traditional behavior of showing the top 10 machine HighScores is preserved.

The idea of showing both machine and personal HighScores as a means of tracking progress makes sense to me given my experiences playing on heavily-used public SM5 cabinets.

New players frequently improve their own score while struggling to break into an overall leaderboard.  The lack of visible leaderboard progress can be demoralizing and this seeks to alleviate that.

This change may make less sense on privately owned cabinets where only 1 or 2 people play.

I am open to feedback on this.