nudge SubtractiveScore BMT slightly left
If the MeasureCounter is centered in the notefield (currently the default default
behavior), the text for SubtractiveScoring is left-aligned and pushed further off
to the right edge of the notefield.  The idea was to allow more breathing room
between the centered MeasureCounter and SubtractiveScoring.

Previously, SubtractiveScoring was hardcoded to always use _wendy small and
x-positioned in a way that specifically accommodated that font.  Since
SubtractiveScoring now uses the ComboFont, it's now necessary to calculate an
appropriate x-position in a more general way.  This commit does that.

If the player opts to move their MeasureCounter to the right (this was the only
behavior in SL3.95 and the default behavior in SL5 until somewhat recently), the
text for SubtractiveScoring is horizontally-center-aligned and has an x-pos that
"mirrors" the MeasureCounter.  This is unchanged.