simplify code for ScreenSelectStyle
I defined a table of which panels to color, per-game, and used it in a loop.

I also added a zoom(0) to the non-chosen choices screen transition.  So, if the
player chooses "2 Players", then the ActorFrames for "1 Player" and "Double" (and
"Solo" if present) will zoom to 0 before the screen transitions, hopefully
emphasizing which choice was chosen.

Finally, this makes use of finishtweening() to prevent potential tween
overflow.