let MenuTimer fade in on SelectMusic
While working on the previous commit, I noticed how ScreenSelectMusic's MenuTimer
was instantly visible, as opposed to the other text in the header which tweened in
its alpha value from 0 to 1.

This commit attempts to change this so that SSM's MenuTimer does fade in, leaving
other MenuTimers prior to SSM to not fade in (too visually distracting).  This
works, ... but there are other issues with MenuTimers that remain.

MenuTimers are weird in that they (in Simply Love, anyway) jump to "0" as part of
the current screen's OffCommand (usually via player interaction), regardless of how
much time was actually remaining.  It's visually distracting and I'd like to
investigate it eventually.