include game + style icons in SSM an Eval headers
This commit adds two 9-panel pad icons to the header for SelectMusic and
Evaluation.  dance, dance-solo, pump, techno, and side-not-active can be
represented this way. They will update in SelectMusic if a player
latejoins or if the player switches from single to double (or solo).

I started to add these to the header for EvalSummary, but realized a
single set of icons there would be misleading if the player had
switched between single and double within a single game session.

...then I realized that EvalSummary's header's GameMode text is already
misleading if multiple GameModes (for example, both ITG and FA+) were
played during a single session.  Clearly, I need to do some work on EvalSummary
but that'll have to wait until tomorrow.