nudge y-offset of TargetScoreGraph
While testing the previous commit, I noticed that the TargetScoreGraph
was drawing about 2 pixels too high, roughly the height of the white
border around it.

I don't quite feel like tracking down when this issue was introduced,
but I'm guessing it was when I changed the global Border() helper
function.