split GameplayExtras into two OptionRows for 4:3
Similar to f5468912637dd9a7470965952f70abf4b04ee158, I took what was a
single OptionRow and present it as two if StepMania isn't using a
WideScreen aspect ratio.

When the engine detects that the choices being presented will be wider
than allowed by (some math resulting from various metrics you have set),
it collapses all the choices into a single BitmapText actor to prevent
the text from going offscreen.

Not going offscreen is good, but not showing all the presentable choices
at once is also kind of confusing.  So, this.

This strategy works when the OptionRow is of SelectType "SelectMultiple"
(like here) but we can't split OptionRows where you're supposed to pick
a single thing (for example, TargetGrade) into multiple lines.