fix custom HoldJudgment for "None"
Already, the first bug. :D

Since the engine is still handling 99% of the logic for HoldJudgments,
it was attempting to set dropped holds ("HoldNoteScore_LetGo") on the
Sprite to the second frame of the "None" sprite.. which didn't exist
because there was no multiframe texture loaded.

I guess we'll add a 1px by 2px transparent png for None and cross our
fingers.