better support for Center1Player + UpperNPSGraph
We should make an effort to show the score and BPM when possible, even
when Center1Player is enabled.

If the Center1Player SM5 preference is enabled and only 1 player is
joined and UpperNPSGraph is requested...

• center the UpperNPSGraph
• score displays on left if P1, or right if P2
• BPM displays on right if P1, or left if P2

Tested in single, double, 16:9, 4:3, Center1Player on/off.

The code in BPMDisplay is looks very questionable to me now, in 2020,
and should be reviewed soon.