include song title in filename for screenshot
This changes ScreenshotHandler.lua in ScreenEval to include the first 10
characters of the song or course in screenshot filenames.

I'm worried that CJK/Cyrillic/etc. characters in filenames will break
Windows based on bug reports I got while working on early versions of
my Watermelon stepchart for UPS5.  But, it's not easy for me to test
Windows.  I can remove this change (or make it Linux + macOS only) if
needed.

Added lots of new inline comments explaining what is going on as well.