hide DisplayMode OptRow in macOS for now
Using the operator menu to toggle fullscreen via DisplayMode is
currently broken in SM5.1-beta on macOS.  I ... don't know what state
5.0.12 was released in, but I also do not have a ready means of testing
it.

For now, let's hide the DisplayMode OptionRow from Graphics/Sound
Options if macOS is detected.  Trying to toggle fullscreen this way
results in rendering breaking somewhere, requiring the player to quit SM
and manually edit Preferences.ini to set Windowed back to 1.

I'm starting to like this practice of conditionally hiding OptionRows
less and less.  It's better than presenting players with an OptionRow
that will result in a bug 100% of the time, but it's still confusing
when different players see different UIs and don't understand why.