SecondsToHMMSS() for SelectMusic header
Game sessions longer than 1 hour but shorter than 10 hours would
previously display with two digits in the hours segment.  1 hour, 15
minutes, and 45 seconds would display as "01:15:45"

The SM5 engine does not (yet) provide a formatting function for HMMSS,
so this creates one in-theme.  Players would now see "1:15:45"

If SecondsToHMMSS is added to the engine in the future, this Lua is
written to use that instead.  This is my attempt to be both forward- and
backwards-compatible.