profile avatars
This is only half-implemented, but I'm committing what I've got for now.

This is the start of a player avatar system.  It will look for an image
with filename "avatar" in the player's local profile first.  If
something like "avatar.png" or "avatar.jpg" is found here, this is what
will be used.

If no avatar is found there, the system will next look to
/StepMania/Appearance/Avatars/
for an image with a filename that matches the profile's DisplayName.

The DisplayName is what players type in the op.menu screen for managing
local profiles (ScreenMiniMenuContext). It is what appears in the
SelectProfile scroller.  It is different from a high score handle.

If the profile's DisplayName (as it appears on SelectProfile) is
"quietly-turning", and there isn't an "avatar" image found in
quietly-turning's profile directory, this system will check for
/StepMania/Appearance/Avatars/quietly-turning.png

(or quietly-turning.jpg)

This system is visually designed for square avatars.  Non-square avatars
will be squished to fit a 1:1 aspect ratio.

If players want to use the /Appearance/Avatars/ setup, they'll need to
create those directories themselves in the root of their StepMania
directory.  This is an effort to ease eventual transition to SM5.3
which already has such a directory structure in place.

---

There's a lot more to implement, refine, and test, but this was already
9 hours of effort today. D: