slightly darken bg of ManageLocalProfiles MiniMenu
This commit slightly darkens the background when viewing the modal
MiniMenu for managing a specific local profile (Set P1, Set P2, Edit,
Rename, Delete, Merge, etc.).

I'd really been hoping to change the options presented (wanted to remove
choices for "Change To Guest", "Change To Normal", and "Change To Test"
as the engine's use of those terms is different than SL's) but the
MiniMenu choices appear to be hardcoded into
ScreenOptionsManageProfiles.cpp.  The necessary Lua hooks to recreate a
similar, reduced menu in Lua are not available, so this task will have
to wait for now.