reposition score, PeakNPS text in StepStats pane
This tidies the alignment and positioning of the percent score and
PeakNPS BitmapText actors when the StepStats pane is active.  Tested in
4:3, 16:9, 16:10, and 21:9.

Some players using DDR upgrade cabinets with 4:3 monitors had noted that
the PeakNPS text was always drawing off the edge of their monitor due to
CRT issues (pincushion, skew, etc.) outside their control.  If nothing
else, the changes in this commit should nudge PeakNPS in from the screen
boundary a little.

There is a way to adjust overall screen zoom and offset from the
Operator Menu (Menu → Visual Options → Overscan Correction) though it's
kind of hidden and confusing to use.  I'd like to make it easier to use,
but probably won't get to it this release cycle.