show unmapped input events on ScreenTestInput
This should help people during the process of mapping dance pads to game
buttons.

If the input event is mapped to a game button, the pad visualization
will light up like before.  If the input event is not mapped, show some
text saying what the event is being registered as, and note that it is
not mapped to any game button.