redo Operator Menu layout to match sub-screens
Because, honestly, even I could never remember whether Default Fail Type
was in Arcade Options or Advanced Options.

This adds a side-pane to the "operator menu" (ScreenOptionsService) that
shows the first six options available on the next screen.

It handles localization properly as far as I've tested.
It works in 16:9, 16:10, 4:3, and 21:9.

It would be *really* nifty if a future commit enhances this even further
to show what the first six settings are set to rather than listing what
they are.  That's unlikely to happen this release cycle.

This is a very large change all at once, so it's likely something is
broken that I haven't detect yet.  If you are beta-testing, let me know
if you find anything.