here in the darkness
I've renamed what was previously titled "The Rabbit Hole" as "here in
the darkness"

This more appropriately captures my current feelings on it as a piece.

It also actively distinguishes my own work here from the unfortunate
YouTube series that seems to delight in taunting humans suffering from
schizophrenia.  That had never been my intent.  It is unfortunate they
shared a name for a while.