StepStats alignment fixes for Center1Player
This is now the ugliest code in Simply Love.

As the inline comment states:

FIXME: this is out of control and points to the need for a generalized
approach to positioning and scaling actors based on AspectRatio (4:3,
16:10, 16:9, 21:9), Step Stats (drawing or not), NPSGraphAtTop (drawing
or not), Center1Player (see GetNotefieldX() in ./Scripts/SL-Helpers.lua)
and which players are joined.

I'm very much looking forward to adding new things again soon, rather
than endlessly adjusting the multitude of configurations ScreenGameplay
needs to support.