better "Stage i" offset calc w/ centered notefield
Simply checking the value of the Center1Player preference isn't good
enough.  It can be true, but 2 players could be joined.  Or, it could be
false, and the player could be playing dance solo, or techno8 singles,
or etc.

The value of the preference alone does not tell us where the
notefield(s) will be.

I went out of my way to write a GetNotefieldX() helper function, so I
might as well use it.