Add lookaheds the the measure counter
- We add two more BMTs to the measure counter to look at upcoming
breaks/streams
- Additinally this change removes the 2 measure minimum to count as a
stream. We now count even 1 measure streams as stream. This is because
we want the Measure table to include as much information as possible to
make the user experience for the lookahead better. We can always change
MeasureCounter.lua to handle this if we want to add that back in.
- Clean up MeasureCounter.lua a bit