adjust MeasureCounter lookahead opacity
This lowers the opacity of all lookahead counts to distinguish them more
from the one "active" count that remains at full opacity.

The mental/visual parsing challenge feels lower like this to me at
least.  Watching it in action, my visual attention on the MeasureCounter
would be drawn to whatever was brighter, which might be two different
stream counts (active stream and active+2 lookahead).

Since the default behavior is to show the active count + 2 lookaheads,
and since the current parsing behavior is "it's stream or it's rest",
the MeasureCounter will always alternate: stream, rest, stream, rest, ...

If you're streaming, the next number will always be a rest, and the next
next number will always be the stream after that.  If you're resting,
the pattern will be flipped: next is stream, next next is rest.

There's no variation, so visually distinguishing an upcoming break from
an upcoming stream seems less important (to me, anyway, but I don't
play, so 🙃).  I'm opting to prioritize being not-visually-distracting
in a way that makes sense to me, but ymmv.  The code is still in place
and fairly easy to modify (thanks, teejusb!) for those that are into
that.