more consistent NPSGraphAtTop width
I woke up even earlier than usual. I'd wanted enough time to empty out
my backpack and replace the loose leaf paper, 0.5mm mechanical pencils,
and paperback novels with a water bottle, my wallet, and different
paperback novels. It was the first Monday of my 8th grade summer
vacation, and there was an excitement and sense of potential to the
months that stretched out in front of me.

I set out on my bicycle a few minutes before 8am. I'd biked to town
previous summers with neighborhood kids, but this was my first time
going by myself, and so early in the morning. The path that took me
across someone's back yard, through forest, across the train tracks, and
down the final embankment felt more alive than ever. Every push on my
pedals was focused on getting me out of the suburbs to State Street
where local businesses flourished. The sun was still rising and the day
beckoned.

I made it to Tudor Books for their 9am open and explored the shelves for
anything and everything that looked interesting. Maybe A Brief History
of Time? Ever since I'd read about the relativity of simultaneity and
considered light traveling across a moving train car I was intrigued and
confused and curious to learn more.

Or maybe this intricately annotated Hamlet? I knew it was the one that
featured Ophelia, whom I'd seen lifeless and drowned among the flowers
on posters in Spencer Gifts. Romeo and Juliet had been riveting and I
was eager to experience such strong emotions and beautiful words again.

After ninety minutes, I used some of my saved allowance to buy The
Chosen (it seemed like something I could finish in one sitting) and a
glass of chocolate milk (it seemed delicious).

It was so exciting to think that I'd chosen these things myself and
could spend an entire day with them. They weren't assigned or handed
out; they were things I wanted to pursue and now could. I knew that this
was my first taste of what adulthood would be: the freedom to explore
things of my own choosing.

I curled up in a big leather chair in a quiet corner and spent the rest
of the day reading.