clean up SystemLayer credit text for CoinMode_Pay
The moon wasn't something I normally gave much thought to.

Death, impermanence, how terrifying eternity seems, coldness, sorrow,
grief – those were in my mind frequently. Drawing the Moon worked for me
then because it gave those abstract ideas a tangible form.

The moon was something I could see in the sky. It existed outside me,
and I could go outside, look up, and see it most nights.

Death and sorrow and grief are things that exist, but...

Well, they are more near, but also more far. I'll encounter them on
Earth, but I can't point at them like I can the moon.

Maybe that's why the moon was such an evocative symbol for my young
mind. I could see it, but it was impractically out of reach. It existed
and I could see if when I wanted to, but it was always far away.

The topics more on my mind didn't have a physical form, but they felt
constantly near. They existed, but not in a way anyone could see.