hitd
Why do you keep reading Nabokov if it hurts so much?

    I want to be okay with the hurt.

To come to terms with it, and be able to put it aside, mostly?

    Yes.  I hope this is possible.

I hope so, too.  I hope you find peace.

    There are a lot of things that don't hurt.  I suppose I should focus
    on them.

There's balance somewhere, surely, though I'm in no position to give
advice on the virtues of balance and focusing on positives.

    Maybe we will both find peace.