minor tidy
after leaving the hallway
a love letter to love letters

eternal sunshine
how do we keep repeating the same mistakes again and again?
why does it never hurt less?
can we still love one another anyway?

le mur
love is hopeless
but that's okay
because we all are

the witness
can logic be applied to solve love like a puzzle?

undertale
is love more than the sum of our choices?

simply love
maybe love is a never-ending work-in-progress.