switch SL's license from MIT to GPLv3
One of the survivors, poking around in the ruins with the point of a
spear, uncovers a singed photo of Richard Stallman. They stare in
silence. "This," one of them finally says, "This is a man who BELIEVED
in something."

---

I'm switching Simply-Love-SM5's license from MIT to GPLv3 because it
aligns most closely with the objectives of this project.

Quite simply, I was not supposed to live past March 31, 2018, so I chose
the MIT the license in 2f68e25f75c5dfe0bd11fa2009bd54745d2a77ba for
SLv4.7 with the idea that it would be a gift to the community and I
wouldn't be alive past that to care what happened.

It didn't play out that way; I'm still here developing for Simply Love
and the StepMania project.  Simply Love v4.8 was very much an Extra Stage.

It seems there have now been 9 additional releases of this project past
the intended final release.  I think it's fair to acknowledge:

1. Simply Love will continue to see more releases going forward.
2. The MIT license that was intended to be the most permissive possible
   now seems less desirable in the context of my efforts being used for
   commercial endeavors and anyone's personal profit.

Regarding external libraries currently in Simply-Love-SM5:

 ./Scripts/utf8.lua is BSD 3-Clause
 ./BGA/_modules/QR Code/qrencode.lua is BSD 3-Clause
 ./Scripts/sha2for51.lua is ???
    it appears to originate from https://stackoverflow.com/a/51561685

There's a single function (TableToString_Recursive) in ./Scripts/06
SL-Utilities.lua that came from http://www.hpelbers.org/lua/print_r
which now 404s.

If you've previously contributed code or assets to Simply Love for free
(I did not commission the work from you for payment) and you don't want
that work to be GPLv3 licensed, please let me know by @ing me on GitHub
and I'll remove it from the project.

If you've previously contributed code or assets to Simply Love that I
paid you for and you don't want that work to be GPLv3 licensed, please @
me on GitHub so we can discuss it.  If you don't have a GitHub account,
I can be reached at quietly.turning at gmail dot com.

It's possible that some code and/or assets in Simply-Love-SM5 right now
that will need to be removed, but I believe I've performed a best-effort
review.  I can address additional things should they come up in the
future.