fix SelectMusic header bug from prev commit
I keep thinking about the library,
my mind still walking up and down the stacks,
even if my body isn’t.

It had the sort of love
I thought was long gone from this world.

Collections of recommended reading
in literature, history, music, art.

Each pile accompanied by a photograph
of the librarian who'd found such delight
in this book and this album,
this film and this painting,
they couldn't wait to share it with you.

They wanted you to know that same joy.

It was exquisite how personal it felt.

Even the computers
felt so pleasantly helpful,
quaint catalog systems
and digital composition demos.

And that was great, really.

It felt like I felt about computers 20 years ago.

Technology was mysterious
but contained so much budding potential.
As though we as humans
were just on the cusp of seeing something else.

It would surely be good,
just beyond, just over that peak,
still slightly out of reach.

There was so much innocent optimism.
So much imagination, and camaraderie,
and hope that humanity would be brought closer together
by the act of sharing knowledge with one another.

I guess I can't go back to that time

before Facebook data mining,
the rise of fascism on Twitter,
and the rush to run studies
on how to more effectively train humans
to kill other humans
through carefully
crafted
algorithms.

But the stacks held me for a while,
retaining me in the present
just long enough, as though to say,
"it is still possible to feel okay."