Fix measure counter in course mode
The measure counter is hidden at the end of the song when there are no
more streams, but never made visible again when the next song starts.
This effectively disables the measure counter after the first song in a
course.

Let's change that by resetting the visibility when the song changes.

Fixes https://github.com/Simply-Love/Simply-Love-SM5/issues/217