Add GlobbalOffsetSeconds to Graphics/Sound options
I believe that adding this to to the options menu and providing a
description and recommendation can help new players to correctly sync
their setup. I forget every time whether negative or positive values are
required to make the arrows later, so this helps me to remember too.

I called the option "Global Offset", because that's what it's called in
Preferences.ini and that's the term experienced players are familiar
with. "Audio Offset" or "Audio Delay" might have been more descriptive,
but oh well...