fix HoldsMinesRolls BMT layout in StepStats
This addresses a small bug introduced in
5de3718984da287b16970e65cd1ae3974f208557 with the way BitmapText actors
for Holds, Mines, and Rolls values were laid out.

The BMTs should be laid out from left to right like:
  player_performance slash possible

Following 5de3718984da287b16970e65cd1ae3974f208557, PLAYER_1's layout
was:
  possible slash player_performance

This addresses that bug so that the layouts match and are correct for
both players.