make ScreenEval's ScreenshotHandler UTF-8 aware
Issue 279 on GitHub reported:

---
Screenshots created by Simply Love apparently contain the first 10 bytes
of the song title, even if the 10-byte boundary falls in the middle of a
UTF-8 character. This can result in invalid UTF-8 sequences in the
resulting filename.

For example:

Neko Neko☆Super Fever Night

gets shortened to:

Neko Neko�

because the "☆" character is 3 bytes long in UTF-8.
---

This commit resolves that issue by replacing a call to Lua 5.1's native
string.sub() with string.utf8sub() which is made available from
./Scripts/utf8.lua

The example provided in the bug report will now be shortened to:
Neko Neko☆